move: 요일 관련 코드들 위치 이동
