+++ src/constants/day.socialWorker.ts
@@ -1,0 +1,11 @@
+export const DAY_KO_TO_EN = {
+  월: 'MONDAY',
+  화: 'TUESDAY',
+  수: 'WEDNESDAY',
+  목: 'THURSDAY',
+  금: 'FRIDAY',
+  토: 'SATURDAY',
+  일: 'SUNDAY',
+} as const;
+
+export const DAY_LABELS = ['월', '화', '수', '목', '금', '토', '일'] as const;
