+++ src/constants/payType.socialWorker.ts
@@ -11,13 +11,3 @@ export const PAY_CODE_TO_LABEL = {
   MONTH: '월급',
   YEAR: '연봉',
 } as const;
-
-export const DAY_KO_TO_EN = {
-  월: 'MONDAY',
-  화: 'TUESDAY',
-  수: 'WEDNESDAY',
-  목: 'THURSDAY',
-  금: 'FRIDAY',
-  토: 'SATURDAY',
-  일: 'SUNDAY',
-} as const;
